move admin tests to user tests
